refactor: test utils extraction
