+++ packages/dependencies/src/getDependents.test.ts
@@ -1,29 +1,10 @@
 import { join, resolve } from 'path'
 
-import { getPluginConfiguration } from '@yarnpkg/cli'
-import { Configuration, Project } from '@yarnpkg/core'
-
-import { YarnContext } from 'monodeploy-types'
+import { setupContext } from '../../../testUtils'
 
 import { getDependents } from '.'
 
 const cwd = resolve('./example-monorepo')
-
-async function setupContext(cwd: string): Promise<YarnContext> {
-    const configuration = await Configuration.find(
-        cwd,
-        getPluginConfiguration(),
-    )
-    const { project, workspace } = await Project.find(configuration, cwd)
-
-    const context: YarnContext = {
-        configuration,
-        project,
-        workspace,
-    }
-
-    return context
-}
 
 const defaultMonodeployConfig = {
     cwd,
