+++ packages/versions/src/getLatestPackageTags.test.ts
@@ -2,7 +2,7 @@ import { join, resolve } from 'path'
 
 import * as npm from '@yarnpkg/plugin-npm'
 
-import { setupContext } from './test_utils'
+import { setupContext } from '../../../testUtils'
 
 import { getLatestPackageTags } from '.'
