+++ testUtils/index.ts
@@ -1,0 +1,2 @@
+export * from './git'
+export * from './misc'
