+++ testUtils/misc.ts
@@ -1,0 +1,25 @@
+import { getPluginConfiguration } from '@yarnpkg/cli'
+import { Configuration, Project } from '@yarnpkg/core'
+import { PortablePath } from '@yarnpkg/fslib'
+
+import { YarnContext } from 'monodeploy-types'
+
+export async function setupContext(cwd: PortablePath): Promise<YarnContext> {
+    const configuration = await Configuration.find(
+        cwd,
+        getPluginConfiguration(),
+    )
+    const { project, workspace } = await Project.find(configuration, cwd)
+
+    if (!workspace) {
+        throw Error('Invalid CWD')
+    }
+
+    const context: YarnContext = {
+        configuration,
+        project,
+        workspace,
+    }
+
+    return context
+}
